OM-DSL: move classes to simplify packages.
